# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "365 Proxy",
        "description": "Handles all traffic between MS 365 and client applications",
        "version": "1.0.0"
    },
    "host": "365proxy.azurewebsites.us",
    "schemes": [
        "https"
    ],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
        "/accmobile/allUsers": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "All Users",
                "description": "Returns an array of all users",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "An array of user email addresses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "user": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accmobile/allIncidents": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "All Incidents",
                "description": "Returns an array of all incidents",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "An array of incident accounts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    },
                                    "link": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "itemId": {
                                        "type": "number"
                                    },
                                    "coords": {
                                        "type": "string"
                                    },
                                    "reasonForVisit": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "ownersLastName": {
                                        "type": "string"
                                    },
                                    "ownersFirstName": {
                                        "type": "string"
                                    },
                                    "ownersTelephoneNumber": {
                                        "type": "string"
                                    },
                                    "pghCode": {
                                        "type": "string"
                                    },
                                    "citationNumber": {
                                        "type": "string"
                                    },
                                    "comments": {
                                        "type": "string"
                                    },
                                    "callOrigin": {
                                        "type": "string"
                                    },
                                    "submittedBy": {
                                        "type": "string"
                                    },
                                    "modifiedBy": {
                                        "type": "string"
                                    },
                                    "officerInitials": {
                                        "type": "string"
                                    },
                                    "open": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accmobile/selectIncident": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "Single incident",
                "description": "Returns an array with a single incident object",
                "parameters": [{
                    "name": "AdvisoryID",
                    "in": "query",
                    "description": "UUID of incident",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "An array of with a single incident object",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    },
                                    "link": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "itemId": {
                                        "type": "number"
                                    },
                                    "coords": {
                                        "type": "string"
                                    },
                                    "reasonForVisit": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "ownersLastName": {
                                        "type": "string"
                                    },
                                    "ownersFirstName": {
                                        "type": "string"
                                    },
                                    "ownersTelephoneNumber": {
                                        "type": "string"
                                    },
                                    "pghCode": {
                                        "type": "string"
                                    },
                                    "citationNumber": {
                                        "type": "string"
                                    },
                                    "comments": {
                                        "type": "string"
                                    },
                                    "callOrigin": {
                                        "type": "string"
                                    },
                                    "submittedBy": {
                                        "type": "string"
                                    },
                                    "modifiedBy": {
                                        "type": "string"
                                    },
                                    "officerInitials": {
                                        "type": "string"
                                    },
                                    "open": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accmobile/selectAnimals": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "Animals per incident",
                "description": "Returns an array with all animals associated with incident",
                "parameters": [{
                    "name": "AdvisoryID",
                    "in": "query",
                    "description": "UUID of incident",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "An array of animals records",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "itemID": {
                                        "type": "number"
                                    },
                                    "incidentID": {
                                        "type": "string"
                                    },
                                    "animalName": {
                                        "type": "string"
                                    },
                                    "animalType": {
                                        "type": "string"
                                    },
                                    "animalBreed": {
                                        "type": "string"
                                    },
                                    "animalCoat": {
                                        "type": "string"
                                    },
                                    "animalSex": {
                                        "type": "string"
                                    },
                                    "animalAge": {
                                        "type": "string"
                                    },
                                    "LicenseNo": {
                                        "type": "string"
                                    },
                                    "LicenseYear": {
                                        "type": "string"
                                    },
                                    "RabbiesVacNo": {
                                        "type": "string"
                                    },
                                    "RabbiesVacExp": {
                                        "type": "string"
                                    },
                                    "Vet": {
                                        "type": "string"
                                    },
                                    "Comments": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accmobile/animalBreeds": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "All animal breeds",
                "description": "Array of animal breeds",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Array of animal breeds",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "breed": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accmobile/animalCoats": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "All animal coats",
                "description": "Array of animal coats",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Array of animal coats",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "coat": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accmobile/vets": {
            "get": {
                "tags": [
                    "ACC Mobile"
                ],
                "security": [{
                    "Bearer": []
                }],
                "summary": "All veterinarians",
                "description": "Array of vets",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Array of vets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "vet": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access Denied"
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}